test: split `sync_pods_suite_test` into separate core + replicant suites
